simase/requirements.txt
Flask
scikit-image
scikit-learn
numpy
Pillow
joblib
streamlit
